move hover to utils
make ExtensionContext variable a global, but all call sites are not updated
 yet

move flag node generation to its own util
