+++ src/providers/flagsView.ts
@@ -3,12 +3,12 @@ import { FeatureFlag, FlagConfiguration, PatchComment } from '../models';
 import { LaunchDarklyAPI } from '../api';
 import { Configuration } from '../configuration';
 import { FlagStore } from '../flagStore';
-import { generateHoverString } from './hover';
-import * as path from 'path';
 import { debounce, map } from 'lodash';
 import { FlagAliases } from './codeRefs';
 import checkExistingCommand from '../utils/common';
 import { Command } from 'vscode';
+import { FlagNode, FlagParentNode, flagToValues } from '../utils/FlagNode';
+import { generateHoverString } from '../utils/hover';
 
 const COLLAPSED = vscode.TreeItemCollapsibleState.Collapsed;
 const NON_COLLAPSED = vscode.TreeItemCollapsibleState.None;
@@ -110,7 +110,7 @@ export class LaunchDarklyTreeViewProvider implements vscode.TreeDataProvider<Fla
 			} else {
 				const updatedFlag = await this.flagStore.getFeatureFlag(element.flagKey);
 				const updatedIdx = this.flagNodes.findIndex((v) => v.flagKey === element.flagKey);
-				const newFlag = await this.flagToValues(updatedFlag.flag, updatedFlag.config, element as FlagParentNode);
+				const newFlag = await flagToValues(updatedFlag.flag, updatedFlag.config, this.config, this.aliases, element as FlagParentNode);
 				this.flagNodes[updatedIdx] = newFlag;
 				return newFlag.children;
 			}
@@ -282,7 +282,7 @@ export class LaunchDarklyTreeViewProvider implements vscode.TreeDataProvider<Fla
 				flagKeys.map((key) => {
 					this.flagStore.getFeatureFlag(key).then((updatedFlag) => {
 						const updatedIdx = this.flagNodes.findIndex((v) => v.flagKey === key);
-						this.flagToValues(updatedFlag.flag, updatedFlag.config).then((newFlagValue) => {
+						flagToValues(updatedFlag.flag, updatedFlag.config, this.config, this.aliases).then((newFlagValue) => {
 							this.flagNodes[updatedIdx] = newFlagValue;
 						});
 					});
@@ -326,29 +326,6 @@ export class LaunchDarklyTreeViewProvider implements vscode.TreeDataProvider<Fla
 		}
 	}
 
-	private flagFactory({
-		label = '',
-		collapsed = NON_COLLAPSED,
-		children = [],
-		ctxValue = '',
-		uri = '',
-		flagKey = '',
-		flagParentName = '',
-		flagVersion = 0,
-	}) {
-		return flagNodeFactory({
-			ctx: this.ctx,
-			label: label,
-			collapsed: collapsed,
-			children: children,
-			ctxValue: ctxValue,
-			uri: uri,
-			flagKey: flagKey,
-			flagParentName: flagParentName,
-			flagVersion: flagVersion,
-		});
-	}
-
 	private async flagToParent(flag: FeatureFlag, env: FlagConfiguration = null): Promise<FlagParentNode> {
 		let envConfig;
 		if (env !== null) {
@@ -378,405 +355,9 @@ export class LaunchDarklyTreeViewProvider implements vscode.TreeDataProvider<Fla
 
 		return item;
 	}
-
-	async flagToValues(
-		flag: FeatureFlag,
-		env: FlagConfiguration = null,
-		flagParent?: FlagParentNode,
-	): Promise<FlagParentNode> {
-		/**
-		 * Get Link for Open Browser and build base flag node.
-		 */
-		let envConfig;
-		if (env !== null) {
-			envConfig = env;
-		} else {
-			try {
-				const env = await this.flagStore.getFeatureFlag(flag.key);
-				envConfig = env.config;
-			} catch (err) {
-				envConfig = new FlagConfiguration();
-			}
-		}
-		let item;
-		if (flagParent) {
-			item = flagParent;
-		} else {
-			item = new FlagParentNode(
-				this.ctx,
-				flag.name,
-				generateHoverString(flag, envConfig, this.config, this.ctx),
-				`${this.config.baseUri}/${this.config.project}/${this.config.env}/features/${flag.key}`,
-				COLLAPSED,
-				[],
-				flag.key,
-				flag._version,
-				envConfig.on,
-				[],
-				'flagParentItem',
-			);
-		}
-
-		/**
-		 * User friendly name for building nested children under parent FlagNode
-		 */
-		const renderedFlagFields = item.children;
-
-		/**
-		 * Build list of tags under "Tags" label
-		 */
-		if (flag.tags.length > 0) {
-			const tags: Array<FlagNode> = flag.tags.map((tag) => this.flagFactory({ label: tag }));
-			renderedFlagFields.push(
-				this.flagFactory({ label: `Tags`, children: tags, collapsed: COLLAPSED, ctxValue: 'tags' }),
-			);
-		}
-		if (this.aliases) {
-			const aliasKeys = this.aliases.getKeys();
-			if (aliasKeys && aliasKeys[flag.key] !== undefined && aliasKeys[flag.key].length > 0) {
-				const aliases: Array<FlagNode> = aliasKeys[flag.key].map((alias) => {
-					const aliasNode = this.flagFactory({ label: alias, collapsed: NON_COLLAPSED, ctxValue: 'flagSearch' });
-					aliasNode.command = {
-						command: 'workbench.action.findInFiles',
-						title: 'Find in Files',
-						arguments: [{ query: alias, triggerSearch: true, matchWholeWord: true, isCaseSensitive: true }],
-					};
-					return aliasNode;
-				});
-				renderedFlagFields.push(
-					this.flagFactory({
-						label: `Aliases`,
-						children: aliases,
-						collapsed: COLLAPSED,
-						ctxValue: 'aliases',
-						flagKey: flag.key,
-					}),
-				);
-			}
-		}
-		/**
-		 * Build view for any Flag Prerequisites
-		 */
-		const prereqs: Array<FlagNode> = [];
-		const flagPrereqs = envConfig.prerequisites;
-		if (typeof flagPrereqs !== 'undefined' && flagPrereqs.length > 0) {
-			flagPrereqs.map((prereq) => {
-				prereqs.push(this.flagFactory({ label: `Flag: ${prereq.key}`, collapsed: NON_COLLAPSED }));
-				prereqs.push(this.flagFactory({ label: `Variation: ${prereq.variation}`, collapsed: NON_COLLAPSED }));
-			});
-			renderedFlagFields.push(
-				this.flagFactory({
-					label: `Prerequisites`,
-					collapsed: prereqs.length > 0 ? COLLAPSED : NON_COLLAPSED,
-					children: prereqs,
-					ctxValue: 'prereq',
-				}),
-			);
-		}
-
-		/**
-		 * Build individual targeting section for variation and targets assigned to each.
-		 */
-		const targets: Array<FlagNode> = [];
-		const flagTargets = envConfig.targets;
-		if (typeof flagTargets !== 'undefined' && flagTargets.length > 0) {
-			flagTargets.map((target) => {
-				targets.push(
-					this.flagFactory({
-						label: `Variation: ${
-							flag.variations[target.variation].name
-								? flag.variations[target.variation].name
-								: flag.variations[target.variation].value
-						}`,
-						ctxValue: 'variation',
-					}),
-					this.flagFactory({ label: `Values: ${target.values}`, ctxValue: 'value' }),
-				);
-			});
-			renderedFlagFields.push(
-				this.flagFactory({
-					label: `Targets`,
-					collapsed: targets.length > 0 ? COLLAPSED : NON_COLLAPSED,
-					children: targets,
-					ctxValue: 'targets',
-				}),
-			);
-		}
-
-		/**
-		 * Build Flag Variations
-		 */
-		const renderedVariations: Array<FlagNode> = [];
-		flag.variations.forEach((variation) => {
-			const variationValue = variation.name
-				? [this.flagFactory({ label: `${JSON.stringify(variation.value)}`, ctxValue: 'value' })]
-				: [];
-			renderedVariations.push(
-				this.flagFactory({
-					label: `${variation.name ? variation.name : JSON.stringify(variation.value)}`,
-					collapsed: variation.name ? COLLAPSED : NON_COLLAPSED,
-					children: variationValue,
-					ctxValue: 'variation',
-				}),
-			);
-
-			if (variation.description) {
-				renderedVariations.push(
-					this.flagFactory({
-						label: `Description: ${variation.description}`,
-						ctxValue: 'description',
-					}),
-				);
-			}
-		});
-		renderedFlagFields.push(
-			this.flagFactory({
-				label: `Variations`,
-				collapsed: COLLAPSED,
-				children: renderedVariations,
-				ctxValue: 'variations',
-			}),
-		);
-
-		/**
-		 * Show number of rules on the Flag
-		 */
-		renderedFlagFields.push(
-			this.flagFactory({
-				label: `Rule count: ${envConfig.rules.length}`,
-				ctxValue: 'flagRules',
-			}),
-		);
-
-		/**
-		 * Build Fallthrough view
-		 */
-		const defaultRule = envConfig.fallthrough;
-		if (defaultRule.variation !== undefined) {
-			const defaultRuleVar = flag.variations[defaultRule.variation];
-			renderedFlagFields.push(
-				this.flagFactory({
-					label: `Default rule: ${defaultRuleVar.name ? defaultRuleVar.name : JSON.stringify(defaultRuleVar.value)}`,
-					ctxValue: 'defaultRule',
-					flagKey: envConfig.key,
-				}),
-			);
-		} else if (defaultRule?.rollout !== undefined) {
-			const fallThroughRollout: Array<FlagNode> = [];
-			if (defaultRule.rollout.bucketBy) {
-				new FlagNode(this.ctx, `BucketBy: ${defaultRule.rollout.bucketBy}`, NON_COLLAPSED);
-			}
-			defaultRule.rollout.variations.map((variation) => {
-				const weight = variation.weight / 1000;
-				const flagVariation = flag.variations[variation.variation];
-				fallThroughRollout.push(
-					this.flagFactory({ label: `Weight: ${weight} % `, ctxValue: 'weight' }),
-					this.flagFactory({
-						label: `Variation: ${flagVariation.name || JSON.stringify(flagVariation.value)}`,
-						ctxValue: 'variation',
-					}),
-				);
-			});
-			renderedFlagFields.push(
-				this.flagFactory({
-					label: `Default Rollout`,
-					collapsed: COLLAPSED,
-					children: fallThroughRollout,
-					ctxValue: 'rollout',
-					flagKey: flag.key,
-				}),
-			);
-		}
-
-		/**
-		 * Build Off Variation view.
-		 * TODO: Render even if undefined since that is valid option.
-		 */
-		if (envConfig.offVariation !== undefined) {
-			const offVar = flag.variations[envConfig.offVariation];
-			if (offVar !== undefined) {
-				renderedFlagFields.push(
-					this.flagFactory({
-						label: `Off variation: ${offVar.name ? offVar.name : JSON.stringify(offVar.value)}`,
-						ctxValue: 'variationOff',
-						flagKey: flag.key,
-					}),
-				);
-			}
-		}
-
-		/**
-		 * Flag Defaults view for new Environments.
-		 */
-		if (flag.defaults !== undefined) {
-			const defOnVar = flag.variations[flag.defaults.onVariation];
-			const defOffVar = flag.variations[flag.defaults.offVariation];
-			renderedFlagFields.push(
-				this.flagFactory({
-					label: `Defaults`,
-					collapsed: COLLAPSED,
-					children: [
-						this.flagFactory({
-							label: `On variation: ${defOnVar.name ? defOnVar.name : JSON.stringify(defOnVar.value)}`,
-							ctxValue: 'variation',
-						}),
-						this.flagFactory({
-							label: `Off variation: ${defOffVar.name ? defOffVar.name : JSON.stringify(defOffVar.value)}`,
-							ctxValue: 'variation',
-						}),
-					],
-				}),
-			);
-		}
-		return item;
-	}
 }
 
-/**
- * Factory function to generate FlagNode
- */
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-export function flagNodeFactory({
-	ctx = null,
-	label = '',
-	collapsed = NON_COLLAPSED,
-	children = [],
-	ctxValue = '',
-	uri = '',
-	flagKey = '',
-	flagParentName = '',
-	flagVersion = 0,
-}): FlagNode {
-	return new FlagNode(ctx, label, collapsed, children, ctxValue, uri, flagKey, flagParentName, flagVersion);
-}
 
-/* eslint-enable @typescript-eslint/explicit-module-boundary-types */
-/**
- * Class representing a Feature flag as vscode TreeItem
- * It is a nested array of FlagNode's to build the view
- */
-export class FlagNode extends vscode.TreeItem {
-	children: Array<unknown> | undefined;
-	contextValue?: string;
-	uri?: string;
-	flagKey?: string;
-	flagParentName?: string;
-	flagVersion: number;
-	command?: vscode.Command;
-	/**
-	 * @param label will be shown in the Treeview
-	 * @param description is shown when hovering over node
-	 * @param collapsibleState is initial state collapsible state
-	 * @param children array of FlagNode's building nested view
-	 * @param contextValue maps to svg resources to show icons
-	 * @param uri URI to build link for Open Browser
-	 * @param flagKey reference to which flag key the treeview item is associated with
-	 * @param flagParentName will match the flag name for top level tree item
-	 */
-	constructor(
-		ctx: vscode.ExtensionContext,
-		public readonly label: string,
-		public readonly collapsibleState: vscode.TreeItemCollapsibleState,
-		children?: Array<unknown>,
-		contextValue?: string,
-		uri?: string,
-		flagKey?: string,
-		flagParentName?: string,
-		flagVersion?: number,
-		command?: vscode.Command,
-	) {
-		super(label, collapsibleState);
-		this.contextValue = contextValue;
-		this.children = children;
-		this.uri = uri;
-		this.flagKey = flagKey;
-		this.flagParentName = flagParentName;
-		this.flagVersion = flagVersion;
-		this.conditionalIcon(ctx, this.contextValue, this.label);
-		this.command = command;
-	}
-
-	// Without this ignore the signature does not match the FlagTree interface
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
-	conditionalIcon(ctx: vscode.ExtensionContext, contextValue: string, label: string, enabled?: boolean): void {
-		/**
-		 * Special handling for open browser. Called in package.json
-		 */
-		if (contextValue == 'flagViewBrowser') {
-			return;
-		} else if (ctx && contextValue) {
-			this.setIcon(ctx, contextValue);
-		}
-	}
-
-	private setIcon(ctx: vscode.ExtensionContext, fileName: string): vscode.ThemeIcon {
-		return (this.iconPath = {
-			id: null,
-			light: ctx.asAbsolutePath(path.join('resources', 'light', fileName + '.svg')),
-			dark: ctx.asAbsolutePath(path.join('resources', 'dark', fileName + '.svg')),
-		});
-	}
-}
-
-export class FlagParentNode extends vscode.TreeItem {
-	children: FlagNode[] | undefined;
-	contextValue?: string;
-	uri?: string;
-	flagKey?: string;
-	flagParentName?: string;
-	flagVersion: number;
-	enabled?: boolean;
-	aliases?: string[];
-
-	/**
-	 * @param label will be shown in the Treeview
-	 * @param tooltip will be shown while hovering over node
-	 * @param uri used when asked to open in browser
-	 * @param collapsibleState is initial state collapsible state
-	 * @param children array of FlagNode's building nested view
-	 * @param flagKey reference to which flag key the treeview item is associated with
-	 */
-	constructor(
-		ctx: vscode.ExtensionContext,
-		public readonly label: string,
-		public readonly tooltip: string | vscode.MarkdownString,
-		uri: string,
-		public collapsibleState: vscode.TreeItemCollapsibleState,
-		children?: FlagNode[],
-		flagKey?: string,
-		flagVersion?: number,
-		enabled?: boolean,
-		aliases?: string[],
-		contextValue?: string,
-	) {
-		super(label, collapsibleState);
-		this.children = children;
-		this.description = flagKey;
-		this.uri = uri;
-		this.flagKey = flagKey;
-		this.flagVersion = flagVersion;
-		this.enabled = enabled;
-		this.contextValue = contextValue;
-		this.conditionalIcon(ctx, this.contextValue, this.enabled);
-		this.aliases = aliases;
-	}
-
-	private conditionalIcon(ctx: vscode.ExtensionContext, contextValue: string, enabled: boolean) {
-		if (ctx && enabled) {
-			this.setIcon(ctx, 'toggleon');
-		} else if (ctx) {
-			this.setIcon(ctx, 'toggleoff');
-		}
-	}
-
-	private setIcon(ctx: vscode.ExtensionContext, fileName: string): vscode.ThemeIcon {
-		return (this.iconPath = {
-			id: null,
-			light: ctx.asAbsolutePath(path.join('resources', 'light', fileName + '.svg')),
-			dark: ctx.asAbsolutePath(path.join('resources', 'dark', fileName + '.svg')),
-		});
-	}
-}
 
 export interface FlagTreeInterface {
 	children: unknown;
